@@ -1,12 +1,15 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
+import ReactDOM from 'react-dom/client';
+import 'semantic-ui-css/semantic.min.css'
 import './index.css';
+import App from './App';
 import reportWebVitals from './reportWebVitals';
- 
-ReactDOM.render(
-    <App />,
-  document.getElementById('root')
+
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement
+);
+root.render(
+  <App />
 );
 
 // If you want to start measuring performance in your app, pass a function
