@@ -2,7 +2,7 @@ namespace API;
 
 public class WeatherForecast
 {
-    public DateTime Date { get; set; }
+    public DateOnly Date { get; set; }
 
     public int TemperatureC { get; set; }
 
