@@ -7,13 +7,13 @@ namespace Persistence
         public static async Task SeedData(DataContext context)
         {
             if (context.Activities.Any()) return;
-
+            
             var activities = new List<Activity>
             {
                 new Activity
                 {
                     Title = "Past Activity 1",
-                    Date = DateTime.Now.AddMonths(-2),
+                    Date = DateTime.UtcNow.AddMonths(-2),
                     Description = "Activity 2 months ago",
                     Category = "drinks",
                     City = "London",
@@ -22,7 +22,7 @@ namespace Persistence
                 new Activity
                 {
                     Title = "Past Activity 2",
-                    Date = DateTime.Now.AddMonths(-1),
+                    Date = DateTime.UtcNow.AddMonths(-1),
                     Description = "Activity 1 month ago",
                     Category = "culture",
                     City = "Paris",
@@ -31,7 +31,7 @@ namespace Persistence
                 new Activity
                 {
                     Title = "Future Activity 1",
-                    Date = DateTime.Now.AddMonths(1),
+                    Date = DateTime.UtcNow.AddMonths(1),
                     Description = "Activity 1 month in future",
                     Category = "culture",
                     City = "London",
@@ -40,7 +40,7 @@ namespace Persistence
                 new Activity
                 {
                     Title = "Future Activity 2",
-                    Date = DateTime.Now.AddMonths(2),
+                    Date = DateTime.UtcNow.AddMonths(2),
                     Description = "Activity 2 months in future",
                     Category = "music",
                     City = "London",
@@ -49,7 +49,7 @@ namespace Persistence
                 new Activity
                 {
                     Title = "Future Activity 3",
-                    Date = DateTime.Now.AddMonths(3),
+                    Date = DateTime.UtcNow.AddMonths(3),
                     Description = "Activity 3 months in future",
                     Category = "drinks",
                     City = "London",
@@ -58,7 +58,7 @@ namespace Persistence
                 new Activity
                 {
                     Title = "Future Activity 4",
-                    Date = DateTime.Now.AddMonths(4),
+                    Date = DateTime.UtcNow.AddMonths(4),
                     Description = "Activity 4 months in future",
                     Category = "drinks",
                     City = "London",
@@ -67,7 +67,7 @@ namespace Persistence
                 new Activity
                 {
                     Title = "Future Activity 5",
-                    Date = DateTime.Now.AddMonths(5),
+                    Date = DateTime.UtcNow.AddMonths(5),
                     Description = "Activity 5 months in future",
                     Category = "drinks",
                     City = "London",
@@ -76,7 +76,7 @@ namespace Persistence
                 new Activity
                 {
                     Title = "Future Activity 6",
-                    Date = DateTime.Now.AddMonths(6),
+                    Date = DateTime.UtcNow.AddMonths(6),
                     Description = "Activity 6 months in future",
                     Category = "music",
                     City = "London",
@@ -85,7 +85,7 @@ namespace Persistence
                 new Activity
                 {
                     Title = "Future Activity 7",
-                    Date = DateTime.Now.AddMonths(7),
+                    Date = DateTime.UtcNow.AddMonths(7),
                     Description = "Activity 2 months ago",
                     Category = "travel",
                     City = "London",
@@ -94,7 +94,7 @@ namespace Persistence
                 new Activity
                 {
                     Title = "Future Activity 8",
-                    Date = DateTime.Now.AddMonths(8),
+                    Date = DateTime.UtcNow.AddMonths(8),
                     Description = "Activity 8 months in future",
                     Category = "film",
                     City = "London",
